feat(mcp): dual-mode revealui-content template (Stage 1 PR-1.2)
First of 13 first-party MCP servers refactored into transport-agnostic
form. Establishes the template the remaining 12 follow in subsequent
PRs.

Pattern: each server splits into two files.
- servers/factories/<name>.ts: exports create<Name>Server() + setCredentials.
  Pure factory, no transport coupling. Lives under servers/factories/ so
  the claim-drift validator's MCP server counter (which reads only the
  top-level servers/ directory) does not mis-count shared infrastructure
  as a standalone server.
- servers/<name>.ts: thin stdio launcher. Imports factory, connects to
  StdioServerTransport, runs as subprocess. Shebang preserved; hypervisor
  spawn() pipeline unchanged.

Consumers needing HTTP hosting (admin, agent runtime) import the factory
directly and wrap it via createNodeStreamableHttpHandler from
@revealui/mcp/streamable-http (PR-1.1).

Shipped:
- packages/mcp/src/servers/factories/revealui-content.ts (new) — owns
  the 5 tools (list_sites, list_content, get_content, list_users,
  site_stats), credential override mechanism, REST-proxy helpers.
- packages/mcp/src/servers/revealui-content.ts — refactored 314->37
  lines, stdio launcher only.
- createRevealuiContentServer + setRevealuiContentCredentials exported
  from @revealui/mcp index.

Testing: 3 new integration cases — factory + createNodeStreamableHttpHandler
+ mock REST API. Verifies capability handshake over HTTP, tool call
proxy with correct auth headers, missing-credentials error path.

MCP total: 164 passing, 5 skipped (up from 161 after PR-1.1).

Follow-ups: 12 more servers port to this pattern — code-validator,
neon, next-devtools, playwright, revealui-email, revealui-memory,
revealui-stripe, stripe, supabase, vercel, vultr-test, _email-provider.

Verification:
- pnpm --filter @revealui/mcp typecheck -> clean
- pnpm --filter @revealui/mcp test -> 164 passed, 5 skipped
- pnpm gate:quick -> PASS (15/15)
